### 2023-02-10 #5144 @editorialbot
[PRE REVIEW]: popkinmocks: mock IFU datacubes for modelling stellar populations and kinematics
**Submitting author:** @prashjet (<a href="http://orcid.org/0000-0003-0010-8129">Prashin Jethwa</a>)
**Repository:** https://github.com/prashjet/popkinmocks
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
