### 2023-03-07 #5225 @editorialbot
[REVIEW]: popkinmocks: mock IFU datacubes for modelling stellar populations and kinematics
**Submitting author:** @prashjet (<a href="http://orcid.org/0000-0003-0010-8129">Prashin Jethwa</a>)
**Repository:** https://github.com/prashjet/popkinmocks
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
