### 2023-04-06 prashjet/popkinmocks#75 @Jammy2211
JOSS Review (Optional): pypi package
This is not required to publish in JOSS (and I have ticked the appropriate box for the review), but at some point in the future I suggest you deploy the package to pypi (https://pypi.org).

Whilst the installation instructions currently work brilliantly and are simple and concise, the benefit of pypi is it means that other projects will easily be able to include `popkinmocks` as a dependency.
