### 2023-04-06 prashjet/popkinmocks#76 @Jammy2211
JOSS Review: Science summary / references
Another minor thing, but I think it would benefit the JOSS paper to have one / two sentences, with references giving some brief highlights of the major scientific results IFU studies have enabled over the past decades.

What you include is at your discretion, but the sort of things I think of are the fast / slow rotator dichotomy from ATLAS3D or other more recent ManGa results.
