### 2023-04-06 prashjet/popkinmocks#77 @Jammy2211
JOSS Review: Community Guidelines / Contributing
The JOSS review has the following tick box:

I may be being blinded, but I can't see anything of the sort on the project.

I would add:

1) A sentence on the main GitHub README.md on how to do this.
2) Something similar on readtheodcs.
3) A CONTRIBUTING.md file (e.g. https://github.com/Jammy2211/PyAutoLens/blob/main/CONTRIBUTING.md)

This is not a JOSS requirement, but I would also recommend adding a CODE_OF_CONDUCT.md (feel free to read and adapt this one: https://github.com/Jammy2211/PyAutoLens/blob/main/CODE_OF_CONDUCT.md)
