### 2023-04-06 #5225 @Jammy2211
All done, I've put up 3 issues on the GitHub which should be straight forward enough to address.

Overall, `popkinmocks` is an awesome little package with clear docs, easy-to-follow examples and it serves a need and purpose not currently filled elsewhere in the literature.

You may hear from me one day about simulating IFU observations of strong lenses with the tool!
