### 2023-04-18 prashjet/popkinmocks#78 @apoudel2014
JOSS Review (Optional): User Guide
I considered leaving this as a comment, but it seems quite obvious. However, I wanted to bring it to your attention just in case it's deemed necessary to make any changes. If not, it's fine to leave it as is. It would be helpful to provide an explanation of the acronym VD (in LOSVD) when it is first introduced in the background section of the user guide, to ensure clarity for readers.
