### 2023-04-27 54c24a8
implement L moments
