+++ popkinmocks/components/base.py
@@ -1,5 +1,5 @@
 import numpy as np
-from scipy import interpolate, special
+from scipy import interpolate, special, stats
 import copy
 from functools import partial
 import os
@@ -332,6 +332,65 @@ class Component(object):
         variance = self.get_central_moment(which_dist, 2, light_weighted=light_weighted)
         return variance
 
+    def get_dispersion(self, which_dist, light_weighted=False):
+        """Get standard deviation of a 1D distribution
+
+        See full docstring of `self.get_mean` for restrictions on `which_dist`
+        and meaning of output.
+
+        Args:
+            which_dist (string): which distribution to take SD of.
+            light_weighted (bool): whether to return light-weighted (True) or
+                mass-weighted (False) quantity
+
+        Returns:
+            float/array: the sd or conditional sd
+
+        """
+        variance = self.get_central_moment(which_dist, 2, light_weighted=light_weighted)
+        sd = variance**0.5
+        return sd
+
+    def get_normalised_central_moment(self, which_dist, j, light_weighted=False):
+        """Get normalised central moment of a 1D distribution
+
+        (Normalised central moment)(j) = (central moment)(j)/dispersion**j 
+
+        Args:
+            which_dist (string): which distribution to take normalised central 
+                moment of.
+            j (int) : moment order
+            light_weighted (bool): whether to return light-weighted (True) or
+                mass-weighted (False) quantity
+
+        Returns:
+            float/array: the normalised central (/conditional?) moment
+
+        """
+        central_moment = self.get_central_moment(
+            which_dist, j, light_weighted=light_weighted
+            )
+        dispersion = self.get_dispersion(
+            which_dist, light_weighted=light_weighted
+            )
+        normalised_central_moment = central_moment/dispersion**j
+        return normalised_central_moment
+    
+    def get_normalised_central_moment_normal(self, j):
+        if j%2 == 0:
+            nrm_mom = special.factorial2(j-1)
+        else:
+            nrm_mom = 0.
+        return nrm_mom
+
+    def get_excess_central_moment(self, which_dist, j, light_weighted=False):
+        nrm_central_moment = self.get_normalised_central_moment(
+            which_dist, j, light_weighted=light_weighted
+            )
+        nrm_moment_normal = self.get_normalised_central_moment_normal(j)
+        excess_central_moment = nrm_central_moment - nrm_moment_normal
+        return excess_central_moment
+
     def get_skewness(self, which_dist, light_weighted=False):
         """Get skewness of a 1D distribution
 
@@ -347,9 +406,9 @@ class Component(object):
             float/array: the variance or conditional skewness
 
         """
-        mu_3 = self.get_central_moment(which_dist, 3, light_weighted=light_weighted)
-        variance = self.get_variance(which_dist, light_weighted=light_weighted)
-        skewness = mu_3 / variance**1.5
+        skewness = self.get_normalised_central_moment(
+            which_dist, 3, light_weighted=light_weighted
+            )
         return skewness
 
     def get_kurtosis(self, which_dist, light_weighted=False):
@@ -367,9 +426,9 @@ class Component(object):
             float/array: the kurtosis or conditional kurtosis
 
         """
-        mu_4 = self.get_central_moment(which_dist, 4, light_weighted=light_weighted)
-        variance = self.get_variance(which_dist, light_weighted=light_weighted)
-        kurtosis = mu_4 / variance**2.0
+        kurtosis = self.get_normalised_central_moment(
[…]
