### 2023-04-27 d1b1434
Science summary / references
closes #76
