+++ joss/paper.bib
@@ -112,3 +112,72 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2021MNRAS.505..991C},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{Emsellem_2007,
+       author = {{Emsellem}, Eric and {Cappellari}, Michele and {Krajnovi{\'c}}, Davor and {van de Ven}, Glenn and {Bacon}, R. and {Bureau}, M. and {Davies}, Roger L. and {de Zeeuw}, P.~T. and {Falc{\'o}n-Barroso}, Jes{\'u}s and {Kuntschner}, Harald and {McDermid}, Richard and {Peletier}, Reynier F. and {Sarzi}, Marc},
+        title = "{The SAURON project - IX. A kinematic classification for early-type galaxies}",
+      journal = {\mnras},
+     keywords = {galaxies: elliptical and lenticular, cD, galaxies: evolution, galaxies: formation, galaxies: kinematics and dynamics, galaxies: structure, Astrophysics},
+         year = 2007,
+        month = aug,
+       volume = {379},
+       number = {2},
+        pages = {401-417},
+          doi = {10.1111/j.1365-2966.2007.11752.x},
+archivePrefix = {arXiv},
+       eprint = {astro-ph/0703531},
+ primaryClass = {astro-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2007MNRAS.379..401E},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@article{Emsellem_2011,
+	doi = {10.1111/j.1365-2966.2011.18496.x},
+	url = {https://doi.org/10.1111%2Fj.1365-2966.2011.18496.x},
+	year = 2011,
+	month = {may},
+	publisher = {Oxford University Press ({OUP})},
+	volume = {414},
+	number = {2},
+	pages = {888--912},
+	author = {Eric Emsellem and Michele Cappellari and Davor Krajnovi{\'{c}
+} and Katherine Alatalo and Leo Blitz and Maxime Bois and Fr{\'{e}}d{\'{e}}ric Bournaud and Martin Bureau and Roger L. Davies and Timothy A. Davis and P. T. de Zeeuw and Sadegh Khochfar and Harald Kuntschner and Pierre-Yves Lablanche and Richard M. McDermid and Raffaella Morganti and Thorsten Naab and Tom Oosterloo and Marc Sarzi and Nicholas Scott and Paolo Serra and Glenn van de Ven and Anne-Marie Weijmans and Lisa M. Young},
+	title = {The {ATLAS}3D project - {III}. A census of the stellar angular momentum within the effective radius of early-type galaxies: unveiling the distribution of fast and slow rotators},
+	journal = {Monthly Notices of the Royal Astronomical Society}
+}
+
+@ARTICLE{perez_2013,
+       author = {{P{\'e}rez}, E. and {Cid Fernandes}, R. and {Gonz{\'a}lez Delgado}, R.~M. and {Garc{\'\i}a-Benito}, R. and {S{\'a}nchez}, S.~F. and {Husemann}, B. and {Mast}, D. and {Rod{\'o}n}, J.~R. and {Kupko}, D. and {Backsmann}, N. and {de Amorim}, A.~L. and {van de Ven}, G. and {Walcher}, J. and {Wisotzki}, L. and {Cortijo-Ferrero}, C. and {CALIFA Collaboration}},
+        title = "{The Evolution of Galaxies Resolved in Space and Time: A View of Inside-out Growth from the CALIFA Survey}",
+      journal = {\apjl},
[…]
