### 2023-04-27 4f1a266
Conribution guide and code of conduct
closes #77
