### 2023-04-27 9e76a5e
define LOSVD
closes #78
