### 2023-04-27 prashjet/popkinmocks#75 @prashjet
> the benefit of pypi is it means that other projects will easily be able to include `popkinmocks` as a dependency.

a good reason to do so! I will return to this in the future for sure.
