### 2023-04-27 prashjet/popkinmocks#77 @prashjet
Yes, thanks for spotting this: a big oversight on my part! I took your contribution guide with a few adjustments, but searched around for a slightly shorter code of conduct.
