### 2023-04-27 prashjet/popkinmocks#78 @prashjet
Super important to define all acronyms, thanks for spotting this.
