### 2023-04-27 321d9ca
Respond to JOSS review (#79)
* Science summary / references
* Conribution guide and code of conduct
* define LOSVD

closes #78
closes #76
closes #77
