+++ joss/paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 # Summary
 
-Integral Field Units (IFUs) are a type of detector which measure both spatial and spectral information. These detectors output 3D data-products known as datacubes: 2D images with a spectrum associated with each image pixel. Beyond the very local Universe where individual stars can be resolved, IFU datacubes are our most information-rich datasets of galaxies, and large samples of these have been observed [e.g. @SAMI_Croom2021].
+Integral Field Units (IFUs) are a type of detector which measure both spatial and spectral information. These detectors output 3D data-products known as datacubes: 2D images with a spectrum associated with each image pixel. Beyond the very local Universe where individual stars can be resolved, IFU datacubes are our most information-rich datasets of galaxies, and large samples of these have been observed [e.g. @SAMI_Croom2021]. Major results from galaxy IFU studies include the discovery of the fast versus slow rotator dichotomy [@Emsellem_2011], evidence for inside-out galaxy growth [@perez_2013], and the ubiquity of warm orbits across galaxy types [@zhu_2018].
 
 Galaxies are built from stars, gas, dust and dark matter. Stars are an especially useful observational probe as many physical properties can be inferred from their spectra. _Stellar populations_ are intrinsic properties such as a star's age or chemical composition, which affect the strengths of spectral absorption lines. _Stellar kinematics_ are the positions and velocities of stars within the galaxy. The component of velocity along the line-of-sight produces Doppler shifting of absorption lines. In the past, stellar populations and kinematics of external galaxies have been modelled separately; often thought of as distinct subfields. Recent work has demonstrated the power of moving beyond this dichotomy to study galaxy evolution [e.g. @Poci2019]. `popkinmocks` is software to create mock observations of IFU datacubes of galaxy stellar light for the era of combined population-kinematic analyses.
 
