### 2023-04-27 #5225 @prashjet
Hi all - many thanks for the positive reviews and useful suggestions. The latest merge to the main branch addresses all your comments (except the pypi suggestion, which I will come back to in the future).

@Jammy2211

> You may hear from me one day about simulating IFU observations of strong lenses with the tool!

I look forward to it!
