### 2023-05-02 be00602
unify calculation method for excess kurtoses
