### 2023-05-03 #5225 @prashjet
> Once those are done, please post the version number and DOI of the archive in this thread. Let me know if you have any questions!

popkinmocks (1.0.1)
https://doi.org/10.5281/zenodo.7890628

Thanks!
