### 2023-05-05 7555c00
Replace journal macros in .bib file
Our TeX template doesn't have the usual astro journal macros defined.

ref: https://github.com/openjournals/joss-reviews/issues/5225
