+++ joss/paper.bib
@@ -80,7 +80,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Zhu2020,
        author = {{Zhu}, Ling and {van de Ven}, Glenn and {Leaman}, Ryan and {Grand}, Robert J.~J. and {Falc{\'o}n-Barroso}, Jes{\'u}s and {Jethwa}, Prashin and {Watkins}, Laura L. and {Mao}, Shude and {Poci}, Adriano and {McDermid}, Richard M. and {Nelson}, Dylan},
         title = "{Disentangling the formation history of galaxies via population-orbit superposition: method validation}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {galaxies: kinematics and dynamics, galaxies: formation, galaxies: structure, methods:numerical, Astrophysics - Astrophysics of Galaxies},
          year = 2020,
         month = aug,
@@ -98,7 +98,7 @@ archivePrefix = {arXiv},
 @ARTICLE{SAMI_Croom2021,
        author = {{Croom}, Scott M. and {Owers}, Matt S. and {Scott}, Nicholas and {Poetrodjojo}, Henry and {Groves}, Brent and {van de Sande}, Jesse and {Barone}, Tania M. and {Cortese}, Luca and {D'Eugenio}, Francesco and {Bland-Hawthorn}, Joss and {Bryant}, Julia and {Oh}, Sree and {Brough}, Sarah and {Agostino}, James and {Casura}, Sarah and {Catinella}, Barbara and {Colless}, Matthew and {Cecil}, Gerald and {Davies}, Roger L. and {Drinkwater}, Michael J. and {Driver}, Simon P. and {Ferreras}, Ignacio and {Foster}, Caroline and {Fraser-McKelvie}, Amelia and {Lawrence}, Jon and {Leslie}, Sarah K. and {Liske}, Jochen and {L{\'o}pez-S{\'a}nchez}, {\'A}ngel R. and {Lorente}, Nuria P.~F. and {McElroy}, Rebecca and {Medling}, Anne M. and {Obreschkow}, Danail and {Richards}, Samuel N. and {Sharp}, Rob and {Sweet}, Sarah M. and {Taranu}, Dan S. and {Taylor}, Edward N. and {Tescari}, Edoardo and {Thomas}, Adam D. and {Tocknell}, James and {Vaughan}, Sam P.},
         title = "{The SAMI Galaxy Survey: the third and final data release}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {galaxies: general, galaxies: kinematics and dynamics, galaxies: star formation, galaxies: stellar content, galaxies: clusters: general, astronomical data bases: surveys, surveys, Astrophysics - Astrophysics of Galaxies},
          year = 2021,
         month = jul,
@@ -116,7 +116,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Emsellem_2007,
        author = {{Emsellem}, Eric and {Cappellari}, Michele and {Krajnovi{\'c}}, Davor and {van de Ven}, Glenn and {Bacon}, R. and {Bureau}, M. and {Davies}, Roger L. and {de Zeeuw}, P.~T. and {Falc{\'o}n-Barroso}, Jes{\'u}s and {Kuntschner}, Harald and {McDermid}, Richard and {Peletier}, Reynier F. and {Sarzi}, Marc},
         title = "{The SAURON project - IX. A kinematic classification for early-type galaxies}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
[…]
