### 2023-05-05 #5225 @dfm
@prashjet — I've opened a tiny PR to your paper fixing some typographical issues in the bibliography. One last thing I'll need from you: Can you update the Zenodo record metadata (there should be an "Edit" button in the top right corner of [this page](https://zenodo.org/record/7890628) when you're logged in) so that the title is "popkinmocks: mock IFU datacubes for modelling stellar populations and kinematics" to match the paper? Then I'm happy to continue with publication!
