### 2023-05-05 62c6bab
Replace journal macros in .bib file (#81)
* Replace journal macros in .bib file

Our TeX template doesn't have the usual astro journal macros defined.

ref: https://github.com/openjournals/joss-reviews/issues/5225

* Removing period from title
